+++ __tests__/mw/logger.ts
@@ -1,16 +1,15 @@
-import { App } from "../../packages/app/src";
-import logger from "../../packages/logger/src";
-import supertest from "supertest";
-
+import { App } from '../../packages/app/src'
+import { logger } from '../../packages/logger/src'
+import supertest from 'supertest'
 
 describe('Logger tests', () => {
   it('should use the logger middleware and log a GET request', (done) => {
-    const originalConsoleLog = console.log;
+    const originalConsoleLog = console.log
 
     console.log = (log) => {
-      expect(log.split(' ')[2]).toBe('GET');
-      console.log = originalConsoleLog;
-      done();
+      expect(log.split(' ')[2]).toBe('GET')
+      console.log = originalConsoleLog
+      done()
     }
 
     const app = new App()
@@ -27,4 +26,4 @@ describe('Logger tests', () => {
         server.close()
       })
   })
-}) 
+})