+++ examples/basic/index.js
@@ -1,5 +1,5 @@
 import { App } from '@tinyhttp/app'
-import logger from '@tinyhttp/logger'
+import { logger } from '@tinyhttp/logger'
 
 const app = new App()